requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
click==8.1.7
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==4.2.16
djangorestframework==3.15.2
drf-yasg==1.21.7
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
minio==7.2.9
packaging==24.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
service-identity==24.1.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
zope.interface==7.0.3
